优化删除分类份方法

+++ src/Handlers/DeleteCategoryHandler.php
@@ -11,8 +11,7 @@ namespace Notadd\Slide\Handlers;
 use Illuminate\Support\Facades\Storage;
 use Notadd\Foundation\Passport\Abstracts\SetHandler as AbstractSetHandler;
 use Notadd\Slide\Models\Category;
-use Notadd\Slide\Models\Group;
-use Notadd\Slide\Models\Picture;
+use Illuminate\Support\Facades\DB;
 
 /**
  * Class ConfigurationHandler.
@@ -26,7 +25,7 @@ class DeleteCategoryHandler extends AbstractSetHandler
      */
     public function execute()
     {
-        $cateId = $this->request->get('category_id');
+        $cateId = $this->request->input('category_id');
 
         $category = Category::where('alias', $cateId)->first();
 
@@ -37,35 +36,41 @@ class DeleteCategoryHandler extends AbstractSetHandler
 
         $groups = $category->groups()->with('pictures')->get();
 
-        foreach($groups as $group)
+        //如果此分类下图集为空，那么直接删除分类
+        if (!$groups)
         {
-            $pictures = $group->pictures()->get();
-            //循环删除某个图集的所有图片
-            foreach($pictures as $picture)
+            $result = $category->delete();
+            if ($result)
             {
-                try{
-                    $picture->delete();
-                }catch (\PDOException $PDOException){
-                    Picture::rollback();
-                    return false;
-                }
-            }
-            //图集图片循环删除完毕后尝试删除当前图集信息
-            try{
-                $group->delete();
-            }catch (\PDOException $PDOException){
-                Group::rollback();
-                return false;
+                return $this->success()->withMessage('删除分类成功');
             }
         }
 
+        foreach($groups as $group)
+        {
+            $pictures = $group->pictures()->get();
+
+            //如果该图集下图片数为零，那么删除该图集
+            if (!count($pictures))
+            {
+                $group->delete();
+            }
+
+            //如果图集不为空，那么循环删除该图集的所有图片
+            foreach($pictures as $picture)
+            {
+                $picture->delete();
+            }
+            //图集图片循环删除完毕后尝试删除当前图集信息
+
+            $group->delete();
+
+        }
         //如果当前分类的搜有图集及其所含图片删除完毕后，尝试删除当前想要删除的分类
-        try{
-            $category->delete();
+        $result = $category->delete();
+        if ($result)
+        {
             $this->success()->withMessage('删除分类成功');
-        }catch (\PDOException $PDOException){
-            Category::rollback();
-            return false;
         }
     }
 }
